Small refactor and fix char creation
FossilOrigin-Name: 557c3d82c18e88651aa121d2acc9d33ded0a4e52b8d79d6fadce7dcca7401667

+++ game/entity/character.h
@@ -25,10 +25,10 @@ struct Character {
         int m_attack_speed;
         int walk_speed;
         int run_speed;
-        int hp;
-        int current_hp;
-        int mp;
-        int current_mp;
+        double hp;
+        double current_hp;
+        double mp;
+        double current_mp;
         int heading;
         int x;
         int y;
